requirements.txt
basicsr==1.4.2
realesrgan==0.3.0
torch==2.2.1
Pillow==10.2.0
numpy==1.26.4
